% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_network.R
\name{as.graph}
\alias{as.graph}
\title{as.graph}
\usage{
as.graph(mat, p1 = 1, p2 = 3, ...)
}
\arguments{
\item{mat}{matrix or data frame describing a network}

\item{p1}{[1] column of first interactor}

\item{p2}{[3] column of second interactor}

\item{...}{further options passed on to igraph::graph()}
}
\value{
igraph object
}
\description{
wrapper to convert a data.frame from RITAN an igraph graph object
}
\examples{
\dontrun{
G <- as.graph(network_list$PID)
}
}
