% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconnectivity_functions.R
\name{cov_undirected}
\alias{cov_undirected}
\title{cov_undirected
function to show the un-directed coverabe between two nodes lists from two networks}
\usage{
cov_undirected(this_nodes1, this_nodes2, this_net1, this_net2)
}
\arguments{
\item{this_nodes1}{list of nodes for first network}

\item{this_nodes2}{list of nodes for second network}

\item{this_net1}{the first network}

\item{this_net2}{the second network}
}
\description{
cov_undirected
function to show the un-directed coverabe between two nodes lists from two networks
}
