% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconnectivity_functions.R
\name{icon_single_within}
\alias{icon_single_within}
\title{icon_single_within
interconnectivity score within a network}
\usage{
icon_single_within(nodes = NULL, net = NULL, s = 10, verbose = TRUE)
}
\arguments{
\item{nodes}{the node labels to use}

\item{net}{the network to use}

\item{s}{[10] the number of repeated random draws to make}

\item{verbose}{[TRUE] if more verbose output should be shown}
}
\description{
icon_single_within
interconnectivity score within a network
}
