% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{resource_reduce}
\alias{resource_reduce}
\title{resource_reduce
Merge terms across resources to reduce the number of redundant and semi-redundant terms}
\usage{
resource_reduce(genesets = NULL, min_overlap = 0.8, verbose = TRUE)
}
\arguments{
\item{genesets}{the input genesets to consider. May be from one or multiple resources.}

\item{min_overlap}{terms that share at least this fraction of genes will be merged}

\item{verbose}{if TRUE, print status and summary output}
}
\value{
the list of terms, after merging to reduce redundnat and semi-redundant terms
}
\description{
resource_reduce
Merge terms across resources to reduce the number of redundant and semi-redundant terms
}
\examples{
require(RITANdata)
r <- resource_reduce( geneset_list$DisGeNet )

}
