% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{show_active_genesets_hist}
\alias{show_active_genesets_hist}
\title{show_active_genesets_hist}
\usage{
show_active_genesets_hist(nbins = 50, ...)
}
\arguments{
\item{nbins}{Number of bins to include in histogram}

\item{...}{further argumants are passed on to plot()}
}
\value{
NULL. The plot is shown.
}
\description{
function to plot distribution of size of active_genesets object
}
\examples{
require(RITANdata)
load_geneset_symbols('GO_slim_generic')
show_active_genesets_hist()

\dontrun{
## Show the distribution of geneset sizes for the default set of geneset resources
load_geneset_symbols()
show_active_genesets_hist()

## Show the distribution of geneset sizes for a specific resource
load_geneset_symbols(gmt="ReactomePathways")
show_active_genesets_hist()

}
}
