% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_network.R
\name{subSIF}
\alias{subSIF}
\title{SubSIF}
\usage{
subSIF(gene.list = NA, Net2Use = c("PID", "TFe", "dPPI", "HPRD", "CCSB",
  "STRING"), minStringScore = 150, minHumanNetScore = 0.4, minScore = 0,
  verbose = TRUE, dedup = FALSE, directed.net = FALSE,
  include_neighbors = FALSE)
}
\arguments{
\item{gene.list}{A list of genes to use. The function will identify edges across resources for or among these genes; identify the induced subnetwork around the gene.list.}

\item{Net2Use}{Name of network resource(s) to use.}

\item{minStringScore}{If STRING is among the Net2Use, only edges of at least the indicated score will be included. Default = 150.}

\item{minHumanNetScore}{If HumanNet is among the Net2Use, only edges of at least the indicated score will be included. Default = 0.4.}

\item{minScore}{Same as above, but used for any other networks where "score" is provided}

\item{verbose}{If TRUE (default), the function will update the user on what it is doing and how many edges are identified for each resource in Net2Use.}

\item{dedup}{If TRUE (Default = FALSE),}

\item{directed.net}{Logical indicating if the network resources should be interpreted as directed.}

\item{include_neighbors}{Logical to include 1st neighbors of "gene.list" (genes not in gene.list, but directly connected to them) in the induced subnetwork.}
}
\value{
Data table describing the induced subnetwork for "gene.list" across the requested resources.
}
\description{
SubSIF
}
\examples{
## Get interactions among a list of genes from the PID: Pathway Interaction Database
require(RITANdata)
myGeneSet <- c('BRCA1','RAD51C','VAV1','HRAS','ABCC1','CYP1B1','CYP3A5')
sif <- subSIF( myGeneSet, Net2Use = 'PID')
print(sif)

\dontrun{
## Get the PPI network induced by genes within myGeneSet
## Use 4 seperate resources, but trim STRING to only include more confident interactions
sif <- subSIF( myGeneSet, c('dPPI','HPRD','CCSB','STRING'), minStringScore = 500 )
}
}

