% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{summary.term_enrichment}
\alias{summary.term_enrichment}
\title{summary.term_enrichment}
\usage{
\method{summary}{term_enrichment}(object, ...)
}
\arguments{
\item{object}{data frame returned by term_enrichment()}

\item{...}{Further arguments are passed on to head()}
}
\value{
the data.frame of top enrichment results
}
\description{
summary.term_enrichment
}
\examples{
require(RITANdata)
e <- term_enrichment( vac1.day0vs31.de.genes, "MSigDB_Hallmarks" )
summary(e, n=3)

}

