% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{summary.term_enrichment_by_subset}
\alias{summary.term_enrichment_by_subset}
\title{summary.term_enrichment_by_subset}
\usage{
\method{summary}{term_enrichment_by_subset}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{data frame returned by term_enrichment_by_subset()}

\item{verbose}{if TRUE (default), print a header describing the data type}

\item{...}{Further arguments are passed on to head()}
}
\value{
the data.frame of top enrichment results
}
\description{
summary.term_enrichment_by_subset
}
\examples{
require(RITANdata)
vac1.de.genes <- list(vac1.day0vs31.de.genes, vac1.day0vs56.de.genes)
names(vac1.de.genes) <- c("Day0vs31", "Day0vs56")
e <- term_enrichment_by_subset(vac1.de.genes, "MSigDB_Hallmarks", q_value_threshold = 0.1 )
summary(e)

}

