% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\docType{data}
\name{vac1.day0vs56.de.genes}
\alias{vac1.day0vs56.de.genes}
\title{This dataset is included as an example in the package:}
\format{An object of class \code{character} of length 471.}
\usage{
vac1.day0vs56.de.genes
}
\value{
differentially expressed genes at 56 days post-vaccination with vaccine1
}
\description{
This dataset is included as an example in the package:
}
\examples{
\dontrun{
 #data("vac1.day0vs56.de.genes")
 te <- term_enrichment(geneset = vac1.day0vs56.de.genes)
}

}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/26755593}
}
\keyword{datasets}
