% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_network.R
\name{write_simple_table}
\alias{write_simple_table}
\title{write_simple_table}
\usage{
write_simple_table(d = NULL, f = NULL, ...)
}
\arguments{
\item{d}{R data object}

\item{f}{file path}

\item{...}{further options passed on to write.table}
}
\value{
invisible (nothing is returned)
}
\description{
This is a simple wrapper around "write.table" that writes a tab-delimited table with column names, no quoting, and no row names.
}
\examples{
\dontrun{
simple wrapper around write.table for writing a tab-delimieted, no row names, tab-seperated file
}
}

