% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{load_all_protein_coding_symbols}
\alias{load_all_protein_coding_symbols}
\title{load_all_protein_coding_symbols}
\usage{
load_all_protein_coding_symbols()
}
\value{
A unique list of gene symbols for protein coding genes according to EnsDb.Hsapiens.v86
}
\description{
The character array returned is, by default, all human protein coding gene symbols. This variable defines the "universe of possible genes" for use in enrichment. Users should load a different "universe" or filter this one down to the most appropriate setting for their current study. For example, if running RNA-Seq, genes are in the universie if they are detected in any sample.
}
