% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{plot.term_enrichment}
\alias{plot.term_enrichment}
\title{plot.term_enrichment}
\usage{
\method{plot}{term_enrichment}(x, min_q = 0.05, max_terms = 25, ...)
}
\arguments{
\item{x}{data frame returned by term_enrichment}

\item{min_q}{Only q-values more significant than this threshold will be plotted}

\item{max_terms}{Up to max_terms will be plotted}

\item{extend_mar}{Extent plot margins since term names can be long strings}

\item{...}{further areguments are sent to plot.}
}
\value{
silent return from plot
}
\description{
plot.term_enrichment
}
\examples{
e <- term_enrichment(vac1.day0vs31.de.genes, term_sources = 'GO_slim_generic')
summary(e)
plot(e, min_q = .1)

}
