% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_enrichment.R
\name{readGMT}
\alias{readGMT}
\title{readGMT}
\usage{
readGMT(f = NA)
}
\arguments{
\item{f}{GMT file name. Please provide a full path if the file is not in the current working directory.}
}
\value{
A list() where the name of each entry is the term (first column of GMT file) and the value is a chr array of genes associated with the term.
}
\description{
Created for simplification of reading .gmt files into RITAN.
}
\examples{
readGMT()

}

